Move WeekLimitSetting to settings
