+++ lib/interface/settings/WeekLimitSetting.module.css
